requirements.txt
streamlit
requests
beautifulsoup4
openai
pydantic
graphviz
pytest
python-dotenv
fastapi
uvicorn
